rename file to axis_format_attributes
