+++ src/plots/cartesian/axis_format_attributes.js
