+++ src/traces/bar/attributes.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var scatterAttrs = require('../scatter/attributes');
-var axisHoverFormat = require('../../plots/cartesian/axis_format').axisHoverFormat;
+var axisHoverFormat = require('../../plots/cartesian/axis_format_attributes').axisHoverFormat;
 var hovertemplateAttrs = require('../../plots/template_attributes').hovertemplateAttrs;
 var texttemplateAttrs = require('../../plots/template_attributes').texttemplateAttrs;
 var colorScaleAttrs = require('../../components/colorscale/attributes');
