+++ src/traces/carpet/axis_attributes.js
@@ -3,7 +3,7 @@
 var fontAttrs = require('../../plots/font_attributes');
 var colorAttrs = require('../../components/color/attributes');
 var axesAttrs = require('../../plots/cartesian/layout_attributes');
-var descriptionWithDates = require('../../plots/cartesian/axis_format').descriptionWithDates;
+var descriptionWithDates = require('../../plots/cartesian/axis_format_attributes').descriptionWithDates;
 var overrideAll = require('../../plot_api/edit_types').overrideAll;
 
 module.exports = {
